Update condition for when to show letterbox
Makes sure we show the letterbox even during the period a window is
requesting to be redrawn because of a layout change (i.e. orientation
change) or for any other reason, if the window's associated surface
is still visible.

This avoids seeing a letterbox flicker when this happens.

Bug: 256052489
Bug: 266635069
Test: Launch the dialer for a phone call or Google wallet app on a large
screen device and make sure the letterbox doesn't flicker on launch.

Change-Id: I7911ddc45c1ff839cc22be09caa5cc7ddc022d45