Rename summaryText where it's not actually related.
summaryText is a property of BigTextStyle notifications, so it's a bit
confusing to reuse the same name for things that are... not that.

Also had to add an m at the beginning of this and related properties
since checkStyle was complaining, and it's now consistent with other
properties.

Test: presubmit
Bug: 269474373
Change-Id: I75d5463fd67afddb4e8f0acc9c12e0a2c8819013
(cherry picked from commit a0cf512354b01205f85986288b3fd09c91986e61)