Track whether we've logged the "intercepted" state of a record.
We didn't have any direct way to determine whether we'd already logged something intercepting (we guessed using the "intercepted" value and whether the record was an update), so we were sometimes double-logging the interception statement.

Further, add additional logging for when the intercepted status is changed for a record. This may happen in the case of re-considering whether a notification should alert after a contacts lookup comes back when DND is on.

Bug: 264918364
Test: manual by confirming zen log after sending a notification updated to break through DND and alert, and various types of notifications that did and did not break through (messages, calls, priority apps)
Test: NotificationManagerServiceTest, ZenModeFilteringTest
Change-Id: Iac37f6af171b05fa3eac61e16be99ae5f36c0b12