Update ASM clear task to only conisder present DisplayArea
If a task is not associated with a DisplayArea, that could throw a NPE.
We now exit early if that is the case.

Extrated the ASM logic to a separate method to make it easier to follow,
and minor updates to logging to match activity starts.

Bug: 270118065
Test: atest CtsAppTestCases
Change-Id: Iabd7c7c3ca503c354014e4fe73d449570cf4ceb5
(cherry picked from commit 2f50b46b96bbf735d28a492131da98c269c190d7)
Merged-In: Iabd7c7c3ca503c354014e4fe73d449570cf4ceb5