Exposes a new CredentialEntry constructor that takes in raw Id
Based on DP1 feedback, if a provider adds > 20 entries we are seeing
a size limit exception. While we are investigating different issues, one
of the potential reasons for more data being passed than is needed is
passing the entire option to the entry, which could contain a no. of strings.
Hence, changing the framework constructor to take in an id only. In
jetpack, we will still take in an option and internally extract the id and
type and pass those to the framework.

Test: CTS tests
Bug: 269501289
Bug: 271298276

Change-Id: Icd6fe6b9a19d137b38785b1ef17c1c13a254e5a6