Inline getVirtualDisplayName into deprecated API
The getVirtualDisplayName is only used in the deprecated virtual display
creation API. The new API uses VirtualDisplayConfig where the client can
set the name.
To make it clear the name as generated is only used in the deprecated
API this commit removes the private method and inlines it in the
deprecated constructor.

Bug: 270334704
Test: VDM Test app, created Virtual display with old and new API
Change-Id: I73726b73aef94276d44a830f83ad74fc764bc1fb
(cherry picked from commit 9ad82520e2641e6fe3b145073202f52f19430bd7)