Compare small and large icons as part of isVisuallyInterruptive()
Like title/text/progress, these fields are visible enough in notifications that we can consider an update to them a "visual difference".

Note that the comparison for Bitmap-based Icons is very approximate and will almost always produce a "true" result. This means that in practice any notification update containing a bitmap will be considered visually different, even if the bitmap is not actually changing. This is not a big difference in behavior, however, because we were already comparing the large icon in the BigPictureStyle with the same caveats and they are likely the same.

Bug: 204367726
Test: atest
Change-Id: I2b58e9201c43fc5a0487b669217ea136e7cefe78