Give fg apps a chance to handle stop before killing its process
With live-tile feature, the removeTask can be called on a task
with resumed activity. And because apps usually save state in
onPause/onSaveInstanceState/onStop, defer the kill a bit until
the app reports it is done.

The timeout is short (1s) because user may want to kill a app
which is no response. If the wait time is too long, the next
launch may still bring the no response app again.

Bug: 184899234
Bug: 213583589
Test: atest RecentTasksTest#testRemovePackageByName

Change-Id: Id69a8ba77da16a13f7a3413bb004fc2206f4b4e8