Revert^2 "Drive SurfaceView visibility on renderthread"
This includes a change to RenderNode which ensures that RenderNode
callbacks are dispatched synchronously in prepareTree(). This is because
the UI thread may be unblocked while replaying drawing commands, and we
don't wait on worker threads for a given frame until after we're done
drawing.

ee17a507326c3084dff0d5892bbcd324cea0f16a

Bug: 274519506
Bug: 269113414
Test: CtsSurfaceControlTests

Change-Id: Iae5c7d096930ec76fd54d9c41a1a479c86d18870