Log unknown package importance as IMPORTANCE_UNSPECIFIED
In the case of packages where for some reason we have package preferences but not any recorded package permissions, log it as such to distinguish from ones that are actually banned by the system (IMPORTANCE_NONE).

This change might help debug cases where packages appear to be banned but we're not sure why.

Bug: 275389727
Test: atest PreferencesHelperTest
Change-Id: Ie22efda4300f63e5f48e08d3148515f04c62643b