DO NOT MERGE Dismiss keyguard when simpin auth'd and...
security method is none. This is mostly to fix the case where we auth
sim pin in the set up wizard and it goes straight to keyguard instead of
the setup wizard activity.

This works with the prevent bypass keyguard flag because the device
should be noe secure in this case.

Fixes: 222446076
Test: turn locked sim on, which opens the sim pin screen. Auth the
screen and observe that keyguard is not shown.
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:3aafb12ff77a386f68473336717dba61fddc21a0)
Merged-In: Ib2d5400e930fa4d6f7f98e46f32a290ba5f558f4
Change-Id: Ib2d5400e930fa4d6f7f98e46f32a290ba5f558f4
Cherry-picking this CL as a security fix

Bug: 222446076
Change-Id: If4360dd6ae2e5f79b43eaf1a29687ac9cc4b6101