Fix KeyguardSecurityContainerControllerTest
Fixing tests that are failing for other build targets.
For orientation, I think the orientation in resources is already defined
as landscape, which is causing the failure in the test.

For the gravity, we have different overlays based on screen size so we
can be so specific with gravity.

Test: Run on ABTD
Fixes: 283210255
Change-Id: Ie341c5d85845098b358f247e6eb8fad05ff0029c
Merged-In: Ie341c5d85845098b358f247e6eb8fad05ff0029c