Close VirtualDevice when CDM association is removed
Close the VirtualDevice and release all resources when the linked CDM
association is removed to ensure VD owners cannot user devices without
the association as the user might have revoked it.

This leaves the virtual device closed and with DEVICE_ID_INVALID,
ensuring new virtual device functionality cannot be invoked.

Bug: 272687853
Test: cts
Change-Id: If03e9236d05ad44abb2c7d9a7aeb26f34ab2279e