Only wait for successful vibration
Currently, the device sleeps for 500ms even if no
vibrator is available, such as on a TV device.
This CL ensures the sleep happens only if the vibration
is happening successfully.

MUST_SLEEP because this is just moving already exiting code.

Test: atest FrameworksServicesTests:VibratorManagerServiceTest
Bug: 276697207
Change-Id: I14182e43ad1fe08c2276be7988f2ac1fb3d79634
Merged-In: I14182e43ad1fe08c2276be7988f2ac1fb3d79634